refactor(backend): Account.id 타입 Int → Long 통일 (account-id-bigint)
레거시(Heroku Postgres) 원본 테이블이라 account.account_id 가 integer 로
덤프돼 와 Account 엔티티만 id: Int 였던 것을 다른 엔티티와 동일하게 Long 으로 통일.

- Flyway V202606052312: account PK + integer FK 8개 테이블(daily_sales_history 포함)
  을 bigint 로 확대. PK+child FK 단일 트랜잭션 동시 ALTER 로 타입 정합 유지.
  (parent_id·monthly_sales_history 는 V58·V202606012103 에서 이미 bigint — 제외)
- 엔티티 4개(Account.id / PPTMaster·AttendanceLog·ProductExpiration.accountId) Long 화
- Repository/Service/Controller/DTO 의 accountId·customerId 타입 Int → Long
- 흩어져 있던 accountId.toInt() 변환 10곳 제거 (출처가 이미 Long)
- 대응 테스트 39개 fixture/mock 타입 정합
- openapi 3종 재생성 (60개 필드 int32 → int64)

integer → bigint widening 이라 기존 값 손실 없이 보존.

+++ backend/src/main/kotlin/com/otoki/powersales/account/dto/response/AdminAccountResponses.kt
@@ -20,7 +20,7 @@ data class AccountListResponse(
 )
 
 data class AccountListItem(
-    val id: Int,
+    val id: Long,
     val externalKey: String?,
     val name: String?,
     val abcType: String?,
@@ -58,7 +58,7 @@ data class AccountListItem(
  * 거래처코드(`externalKey`)는 화면 식별자로 노출하되, `werk*`/`sales_dept_*` 등 내부 SAP 코드는 제외.
  */
 data class AccountDetailResponse(
-    val id: Int,
+    val id: Long,
     val externalKey: String?,
     val name: String?,
     val accountGroup: String?,
@@ -167,7 +167,7 @@ data class AccountDetailResponse(
  * 매칭 실패 시 NULL.
  */
 data class AdminAccountCreateResponse(
-    val id: Int,
+    val id: Long,
     val name: String,
     val accountGroup: String,
     val employeeCode: String,
@@ -196,7 +196,7 @@ data class AdminAccountCreateResponse(
  * #640 [AdminAccountCreateResponse] 와 별도 (필드 범위 차이 — 본 응답은 수정 가능 일반 필드 광범위 포함).
  */
 data class AdminAccountUpdateResponse(
-    val id: Int,
+    val id: Long,
     val name: String?,
     val accountGroup: String?,
     val employeeCode: String?,
