+++ backend/src/main/kotlin/com/otoki/powersales/admin/controller/AdminMonthlySalesDashboardController.kt
@@ -54,7 +54,7 @@ class AdminMonthlySalesDashboardController(
         @RequestParam year: Int,
         @RequestParam month: Int,
         @RequestParam costCenterCodes: List<String>,
-        @RequestParam(required = false) accountIds: List<Int>?,
+        @RequestParam(required = false) accountIds: List<Long>?,
         @RequestParam(required = false) accountGroup: String?,
         @RequestParam(required = false) customerKeyword: String?,
         @RequestParam(required = false, defaultValue = "0") page: Int,
@@ -79,7 +79,7 @@ class AdminMonthlySalesDashboardController(
         @RequestParam year: Int,
         @RequestParam month: Int,
         @RequestParam costCenterCodes: List<String>,
-        @RequestParam(required = false) accountIds: List<Int>?,
+        @RequestParam(required = false) accountIds: List<Long>?,
         @RequestParam(required = false) accountGroup: String?,
         @RequestParam(required = false) customerKeyword: String?,
         @RequestParam(required = false) sort: String?,
@@ -100,7 +100,7 @@ class AdminMonthlySalesDashboardController(
     @GetMapping("/detail/{customerId}")
     fun getDetail(
         @CurrentDataScope scope: DataScope,
-        @PathVariable customerId: Int,
+        @PathVariable customerId: Long,
         @RequestParam year: Int,
         @RequestParam month: Int,
     ): ResponseEntity<ApiResponse<MonthlySalesDashboardDetailResponse>> {
